chore: refactor to ignore errors and show empty state
