+++ packages/extension/src/ui/features/accountActivity/PendingTransactions.tsx
@@ -5,35 +5,9 @@ import { FC } from "react"
 import { Network } from "../../../shared/network"
 import { Token } from "../../../shared/token/type"
 import { Transaction } from "../../../shared/transactions"
-import { BaseWalletAccount } from "../../../shared/wallet.model"
-import { useAppState } from "../../app.state"
 import { openBlockExplorerTransaction } from "../../services/blockExplorer.service"
-import { useAccountTransactions } from "../accounts/accountTransactions.state"
-import { useTokensInNetwork } from "../accountTokens/tokens.state"
-import { useCurrentNetwork } from "../networks/useNetworks"
 import { TransactionListItem } from "./TransactionListItem"
 import { transformTransaction } from "./transform"
-
-interface PendingTransactionsContainerProps {
-  account: BaseWalletAccount
-}
-
-export const PendingTransactionsContainer: FC<
-  PendingTransactionsContainerProps
-> = ({ account }) => {
-  const network = useCurrentNetwork()
-  const { pendingTransactions } = useAccountTransactions(account)
-  const { switcherNetworkId } = useAppState()
-  const tokensByNetwork = useTokensInNetwork(switcherNetworkId)
-  return (
-    <PendingTransactions
-      pendingTransactions={pendingTransactions}
-      network={network}
-      tokensByNetwork={tokensByNetwork}
-      accountAddress={account.address}
-    />
-  )
-}
 
 interface PendingTransactionsProps {
   pendingTransactions: Transaction[]
